+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## main
 * Allow to assign slot to 2D and 3D location indicators.
+* Allow to add slots at runtime.
 
 ## 11.3.0 - 10 April, 2024